+++ NEWS.md
@@ -36,7 +36,7 @@ Language changes
 
 Multi-threading changes
 -----------------------
-
+* There is a new struct `Lockable{T, L<:AbstractLock}` that makes it easy to bundle a resource and its lock together.
 
 Build system changes
 --------------------